Add option to build development version of the gateway

+++ image/build-gateway.sh
@@ -79,9 +79,13 @@ npm install -g yarn
 
 # Install development-specific node modules and prerun webpack
 (cd ${GATEWAY}; rm -rf node_modules; yarn --ignore-scripts; npm rebuild --arch=${ARCH} --target_arch=arm; node_modules/.bin/webpack)
-# Clear out node_modules just in case then build only the modules we'll use in
-# the image, cross compiling any native code.
-(cd ${GATEWAY}; rm -rf node_modules; yarn --ignore-scripts --production; npm rebuild --arch=${ARCH} --target_arch=arm)
+
+echo "DEV_BUILD=${DEV_BUILD}"
+if [ "${DEV_BUILD}" == "0" ]; then
+  # Clear out node_modules just in case then build only the modules we'll use in
+  # the image, cross compiling any native code.
+  (cd ${GATEWAY}; rm -rf node_modules; yarn --ignore-scripts --production; npm rebuild --arch=${ARCH} --target_arch=arm)
+fi
 
 set -x
